Update to current master

+++ support/cas-server-support-redis-ticket-registry/src/main/java/org/apereo/cas/ticket/registry/RedisTicketRegistry.java
@@ -6,12 +6,18 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ScanOptions;
 
+import java.io.IOException;
 import java.util.Collection;
 import java.util.Objects;
+import java.util.Spliterator;
+import java.util.Spliterators;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 /**
  * Key-value ticket registry implementation that stores tickets in redis keyed on the ticket ID.
@@ -23,6 +29,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class RedisTicketRegistry extends AbstractTicketRegistry {
     private static final String CAS_TICKET_PREFIX = "CAS_TICKET:";
+    private static final long SCAN_COUNT = 100L;
 
     private final RedisTemplate<String, Ticket> client;
 
@@ -34,7 +41,9 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
      */
     private static Long getTimeout(final Ticket ticket) {
         val ttl = ticket.getExpirationPolicy().getTimeToLive();
-        if (ttl <= 0) {
+        if (ttl > Integer.MAX_VALUE) {
+            return (long) Integer.MAX_VALUE;
+        } else if (ttl <= 0) {
             return 1L;
         }
         return ttl;
@@ -51,6 +60,10 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
     @Override
     public long deleteAll() {
         val redisKeys = this.client.keys(getPatternTicketRedisKey());
+        if (redisKeys == null) {
+            LOGGER.warn("Unable to locate tickets via redis key");
+            return 0;
+        }
         val size = redisKeys.size();
         this.client.delete(redisKeys);
         return size;
@@ -59,7 +72,7 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
     @Override
     public boolean deleteSingleTicket(final String ticketId) {
         try {
-            val redisKey = getTicketRedisKey(ticketId);
+            val redisKey = getTicketRedisKey(encodeTicketId(ticketId));
             this.client.delete(redisKey);
             return true;
         } catch (final Exception e) {
@@ -72,7 +85,7 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
     public void addTicket(final Ticket ticket) {
         try {
             LOGGER.debug("Adding ticket [{}]", ticket);
-            val redisKey = getTicketRedisKey(ticket.getId());
+            val redisKey = getTicketRedisKey(encodeTicketId(ticket.getId()));
             val encodeTicket = encodeTicket(ticket);
             val timeout = getTimeout(ticket);
             this.client.boundValueOps(redisKey).set(encodeTicket, timeout.longValue(), TimeUnit.SECONDS);
@@ -84,7 +97,7 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
     @Override
     public Ticket getTicket(final String ticketId, final Predicate<Ticket> predicate) {
         try {
-            val redisKey = getTicketRedisKey(ticketId);
+            val redisKey = getTicketRedisKey(encodeTicketId(ticketId));
             val t = this.client.boundValueOps(redisKey).get();
             if (t != null) {
                 val result = decodeTicket(t);
@@ -92,7 +105,7 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
                     return result;
                 }
                 LOGGER.debug("The condition enforced by the predicate [{}] cannot successfully accept/test the ticket id [{}]", ticketId,
-                    predicate.getClass().getSimpleName());
+                        predicate.getClass().getSimpleName());
                 return null;
             }
         } catch (final Exception e) {
@@ -103,7 +116,14 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
 
     @Override
     public Collection<? extends Ticket> getTickets() {
-        return this.client.keys(getPatternTicketRedisKey()).stream()
+        try (val ticketsStream = getTicketsStream()) {
+            return ticketsStream.collect(Collectors.toSet());
+        }
+    }
+
+    @Override
+    public Stream<? extends Ticket> getTicketsStream() {
+        return getKeysStream()
             .map(redisKey -> {
                 val ticket = this.client.boundValueOps(redisKey).get();
                 if (ticket == null) {
@@ -113,8 +133,8 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
                 return ticket;
             })
             .filter(Objects::nonNull)
-            .map(this::decodeTicket)
-            .collect(Collectors.toSet());
+            .map(this::decodeTicket);
+
     }
 
     @Override
@@ -122,7 +142,7 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
         try {
             LOGGER.debug("Updating ticket [{}]", ticket);
             val encodeTicket = this.encodeTicket(ticket);
-            val redisKey = getTicketRedisKey(ticket.getId());
+            val redisKey = getTicketRedisKey(encodeTicketId(ticket.getId()));
             LOGGER.debug("Fetched redis key [{}] for ticket [{}]", redisKey, ticket);
 
             val timeout = getTimeout(ticket);
@@ -132,5 +152,29 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
             LOGGER.error("Failed to update [{}]", ticket, e);
         }
         return null;
+    }
+
+    /**
+     * Get a stream of all CAS-related keys from Redis DB.
+     *
+     * @return stream of all CAS-related keys from Redis DB
+     */
+    private Stream<String> getKeysStream() {
+        val cursor = client.getConnectionFactory().getConnection()
+                .scan(ScanOptions.scanOptions().match(getPatternTicketRedisKey())
+                .count(SCAN_COUNT)
+                .build());
+        return StreamSupport
+            .stream(Spliterators.spliteratorUnknownSize(cursor, Spliterator.ORDERED), false)
+            .map(key -> (String) client.getKeySerializer().deserialize(key))
+            .collect(Collectors.toSet())
+            .stream()
+            .onClose(() -> {
+                try {
+                    cursor.close();
+                } catch (final IOException e) {
+                    LOGGER.error("Could not close Redis connection", e);
+                }
+            });
     }
 }
